Bithumb API Changes Fixes #4 - Wrap IOException
Bithumb API Changes Fixes #4 - Response classes where really for unit tests - removed
