+++ xchange-bithumb/src/test/resources/org/knowm/xchange/bithumb/dto/trade/example-user-transaction.json
@@ -6,7 +6,7 @@
       "transfer_date"     : 1572252297148997,
       "order_currency"    : "BTC",
       "payment_currency"  : "KRW",
-      "units"             : "0.0001",
+      "units"             : "+ 0.0001",
       "price"             : "10000000",
       "amount"            : "1000",
       "fee_currency"      : "KRW",
@@ -19,7 +19,7 @@
       "transfer_date"     : 1569982658530516,
       "order_currency"    : "BTC",
       "payment_currency"  : "KRW",
-      "units"             : "0.0003",
+      "units"             : "- 0.0003",
       "price"             : "9001000",
       "amount"            : "2700",
       "fee_currency"      : "KRW",
